iroiro tsuika

+++ src/ft_utils_0.c
@@ -6,7 +6,7 @@
 /*   By: mayocorn <twitter@mayocornsuki>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/27 08:19:32 by stsunoda          #+#    #+#             */
-/*   Updated: 2022/03/01 11:07:53 by mayocorn         ###   ########.fr       */
+/*   Updated: 2022/03/02 15:09:55 by mayocorn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ int	ft_min(int a, int b)
 
 int	ft_count_digit(int n)
 {
-	size_t	res;
+	int	res;
 
 	res = 0;
 	while (n)
@@ -39,9 +39,9 @@ int	ft_count_digit(int n)
 	return (res);
 }
 
-size_t	ft_count_digit_u(unsigned int n)
+int	ft_count_digit_u(unsigned int n)
 {
-	size_t	res;
+	int	res;
 
 	res = 0;
 	while (n)
@@ -52,9 +52,9 @@ size_t	ft_count_digit_u(unsigned int n)
 	return (res);
 }
 
-size_t	ft_count_digit_h(unsigned long n)
+int	ft_count_digit_h(unsigned long n)
 {
-	size_t	res;
+	int	res;
 
 	res = 0;
 	while (n)
